[MOB-1497] Extract shared migration broadcast-failure sheet component
MigrationSendingView and MigrationNoteSplitView each carried a
byte-for-byte identical ~97-line failureSheetContent/Title/Body/Buttons
group switching on the compliance-critical R14-R17 failureKind routing.
Extracted into MigrationBroadcastFailureSheetView (plain values +
closures, since the two screens' stores have distinct Action types) so
this copy/routing can no longer drift between the two screens. Both
views now instantiate the shared component; visual and behavioral
output is unchanged.
